moar V2 cleanup
